Move `call_parameter_sets` out of `protocols` dir, as it is not a protocol
